bot/requirements.txt
python-telegram-bot==21.6
httpx==0.27.0
supabase==2.5.0
